Updated Supplier service

+++ inventory-supplier-service/src/main/java/com/cg/inventorysupplierservice/dao/DistributorDao.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 @Repository
-public interface DistributorDao extends JpaRepository <Distributor, Integer>{
+public interface DistributorDao extends JpaRepository <Distributor, Long>{
 
 }
