+++ inventory-supplier-service/src/main/java/com/cg/inventorysupplierservice/dao/WarehouseDao.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.cg.inventorysupplierservice.entity.Warehouse;
 
 @Repository
-public interface WarehouseDao extends JpaRepository<Warehouse, Integer> {
+public interface WarehouseDao extends JpaRepository<Warehouse, Long> {
 
+	
 }
